LIME version 5.2.4-530 installable apk
This is the last version of LIME, the input method app for Android.  Due to lack of maintenance of LIME, I place the app in .apk form here under permission of Art Hung, the author of LIME.  You are welcome to download, install and try LIME. The URL of official LIME-IME project is https://github.com/lime-ime/limeime.